You guys stop being assholes, okay?
Codes of conduct are great and necessary today. It sucks that we have
to have a contract that keeps you from acting like a fucking asshole to
folks around you. If you don't like codes of conduct or Contributor
Covenant then go fuck yourself, ideally using this library to go fuck
yourself.